requirements.txt
anthropic>=0.30.0
chromadb>=0.5.0
ollama>=0.2.0
python-telegram-bot>=21.0
openai>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
flask>=3.0.0
litellm>=1.40.0
python-dotenv>=1.0.0
